fix: Refactor logging statements for improved clarity in serialization and deserialization processes
